@@ -7,7 +7,8 @@ public static partial class BufferManager
 	/// Registers object space.
 	/// </summary>
 	/// <typeparam name="TSpace">Type of object space.</typeparam>
-	[ExcludeFromCodeCoverage, MethodImpl(MethodImplOptions.AggressiveInlining)]
+	[ExcludeFromCodeCoverage]
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static void RegisterSpace<TSpace>() where TSpace : struct, IManagedBinaryBuffer<TSpace, Object>
 		=> MetadataManager<Object>.RegisterBufferSpace<TSpace>();
 	/// <summary>
@@ -15,7 +16,8 @@ public static partial class BufferManager
 	/// </summary>
 	/// <typeparam name="T">Type of items in the buffer.</typeparam>
 	/// <typeparam name="TSpace">Type of <typeparamref name="T"/> buffer.</typeparam>
-	[ExcludeFromCodeCoverage, MethodImpl(MethodImplOptions.AggressiveInlining)]
+	[ExcludeFromCodeCoverage]
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static void RegisterSpace<T, TSpace>()
 		where TSpace : struct, IManagedBinaryBuffer<TSpace, T> where T : struct
 		=> MetadataManager<T>.RegisterBufferSpace<TSpace>();
@@ -24,7 +26,8 @@ public static partial class BufferManager
 	/// </summary>
 	/// <typeparam name="T">Type of nullable items in the space.</typeparam>
 	/// <typeparam name="TSpace">Type of <see name="Nullable{T}"/> space.</typeparam>
-	[ExcludeFromCodeCoverage, MethodImpl(MethodImplOptions.AggressiveInlining)]
+	[ExcludeFromCodeCoverage]
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static void RegisterNullableSpace<T, TSpace>()
 		where TSpace : struct, IManagedBinaryBuffer<TSpace, T?> where T : struct
 		=> MetadataManager<T?>.RegisterBufferSpace<TSpace>();
@@ -32,7 +35,8 @@ public static partial class BufferManager
 	/// Registers object space.
 	/// </summary>
 	/// <typeparam name="TSpace">Type of object space.</typeparam>
-	[ExcludeFromCodeCoverage, MethodImpl(MethodImplOptions.AggressiveInlining)]
+	[ExcludeFromCodeCoverage]
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static void RegisterSpace2<TSpace>() where TSpace : struct, IManagedBinaryBuffer<TSpace, Object>
 		=> BufferManager.RegisterSpace<Composite<TSpace, TSpace, Object>>();
 	/// <summary>
@@ -40,7 +44,8 @@ public static partial class BufferManager
 	/// </summary>
 	/// <typeparam name="T">Type of items in the buffer.</typeparam>
 	/// <typeparam name="TSpace">Type of <typeparamref name="T"/> buffer.</typeparam>
-	[ExcludeFromCodeCoverage, MethodImpl(MethodImplOptions.AggressiveInlining)]
+	[ExcludeFromCodeCoverage]
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static void RegisterSpace2<T, TSpace>()
 		where TSpace : struct, IManagedBinaryBuffer<TSpace, T> where T : struct
 		=> BufferManager.RegisterSpace<T, Composite<TSpace, TSpace, T>>();
@@ -49,7 +54,8 @@ public static partial class BufferManager
 	/// </summary>
 	/// <typeparam name="T">Type of nullable items in the space.</typeparam>
 	/// <typeparam name="TSpace">Type of <see name="Nullable{T}"/> space.</typeparam>
-	[ExcludeFromCodeCoverage, MethodImpl(MethodImplOptions.AggressiveInlining)]
+	[ExcludeFromCodeCoverage]
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static void RegisterNullableSpace2<T, TSpace>()
 		where TSpace : struct, IManagedBinaryBuffer<TSpace, T?> where T : struct
 		=> BufferManager.RegisterNullableSpace<T, Composite<TSpace, TSpace, T?>>();
@@ -57,7 +63,8 @@ public static partial class BufferManager
 	/// Registers object space.
 	/// </summary>
 	/// <typeparam name="TSpace">Type of object space.</typeparam>
-	[ExcludeFromCodeCoverage, MethodImpl(MethodImplOptions.AggressiveInlining)]
+	[ExcludeFromCodeCoverage]
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static void RegisterSpace4<TSpace>() where TSpace : struct, IManagedBinaryBuffer<TSpace, Object>
 		=> BufferManager.RegisterSpace2<Composite<TSpace, TSpace, Object>>();
 	/// <summary>
@@ -65,7 +72,8 @@ public static partial class BufferManager
 	/// </summary>
 	/// <typeparam name="T">Type of items in the buffer.</typeparam>
 	/// <typeparam name="TSpace">Type of <typeparamref name="T"/> buffer.</typeparam>
-	[ExcludeFromCodeCoverage, MethodImpl(MethodImplOptions.AggressiveInlining)]
+	[ExcludeFromCodeCoverage]
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static void RegisterSpace4<T, TSpace>()
 		where TSpace : struct, IManagedBinaryBuffer<TSpace, T> where T : struct
 		=> BufferManager.RegisterSpace2<T, Composite<TSpace, TSpace, T>>();
@@ -74,7 +82,8 @@ public static partial class BufferManager
 	/// </summary>
 	/// <typeparam name="T">Type of nullable items in the space.</typeparam>
 	/// <typeparam name="TSpace">Type of <see name="Nullable{T}"/> space.</typeparam>
-	[ExcludeFromCodeCoverage, MethodImpl(MethodImplOptions.AggressiveInlining)]
+	[ExcludeFromCodeCoverage]
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static void RegisterNullableSpace4<T, TSpace>()
 		where TSpace : struct, IManagedBinaryBuffer<TSpace, T?> where T : struct
 		=> BufferManager.RegisterNullableSpace2<T, Composite<TSpace, TSpace, T?>>();
@@ -82,7 +91,8 @@ public static partial class BufferManager
 	/// Registers object space.
 	/// </summary>
 	/// <typeparam name="TSpace">Type of object space.</typeparam>
-	[ExcludeFromCodeCoverage, MethodImpl(MethodImplOptions.AggressiveInlining)]
+	[ExcludeFromCodeCoverage]
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static void RegisterSpace8<TSpace>() where TSpace : struct, IManagedBinaryBuffer<TSpace, Object>
 		=> BufferManager.RegisterSpace4<Composite<TSpace, TSpace, Object>>();
 	/// <summary>
@@ -90,7 +100,8 @@ public static partial class BufferManager
 	/// </summary>
 	/// <typeparam name="T">Type of items in the buffer.</typeparam>
 	/// <typeparam name="TSpace">Type of <typeparamref name="T"/> buffer.</typeparam>
-	[ExcludeFromCodeCoverage, MethodImpl(MethodImplOptions.AggressiveInlining)]
+	[ExcludeFromCodeCoverage]
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static void RegisterSpace8<T, TSpace>()
 		where TSpace : struct, IManagedBinaryBuffer<TSpace, T> where T : struct
 		=> BufferManager.RegisterSpace4<T, Composite<TSpace, TSpace, T>>();
@@ -99,7 +110,8 @@ public static partial class BufferManager
 	/// </summary>
 	/// <typeparam name="T">Type of nullable items in the space.</typeparam>
 	/// <typeparam name="TSpace">Type of <see name="Nullable{T}"/> space.</typeparam>
-	[ExcludeFromCodeCoverage, MethodImpl(MethodImplOptions.AggressiveInlining)]
+	[ExcludeFromCodeCoverage]
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static void RegisterNullableSpace8<T, TSpace>()
 		where TSpace : struct, IManagedBinaryBuffer<TSpace, T?> where T : struct
 		=> BufferManager.RegisterNullableSpace4<T, Composite<TSpace, TSpace, T?>>();
@@ -107,7 +119,8 @@ public static partial class BufferManager
 	/// Registers object space.
 	/// </summary>
 	/// <typeparam name="TSpace">Type of object space.</typeparam>
-	[ExcludeFromCodeCoverage, MethodImpl(MethodImplOptions.AggressiveInlining)]
+	[ExcludeFromCodeCoverage]
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static void RegisterSpace16<TSpace>() where TSpace : struct, IManagedBinaryBuffer<TSpace, Object>
 		=> BufferManager.RegisterSpace8<Composite<TSpace, TSpace, Object>>();
 	/// <summary>
@@ -115,7 +128,8 @@ public static partial class BufferManager
 	/// </summary>
 	/// <typeparam name="T">Type of items in the buffer.</typeparam>
 	/// <typeparam name="TSpace">Type of <typeparamref name="T"/> buffer.</typeparam>
-	[ExcludeFromCodeCoverage, MethodImpl(MethodImplOptions.AggressiveInlining)]
+	[ExcludeFromCodeCoverage]
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static void RegisterSpace16<T, TSpace>()
 		where TSpace : struct, IManagedBinaryBuffer<TSpace, T> where T : struct
 		=> BufferManager.RegisterSpace8<T, Composite<TSpace, TSpace, T>>();
@@ -124,7 +138,8 @@ public static partial class BufferManager
 	/// </summary>
 	/// <typeparam name="T">Type of nullable items in the space.</typeparam>
 	/// <typeparam name="TSpace">Type of <see name="Nullable{T}"/> space.</typeparam>
-	[ExcludeFromCodeCoverage, MethodImpl(MethodImplOptions.AggressiveInlining)]
+	[ExcludeFromCodeCoverage]
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static void RegisterNullableSpace16<T, TSpace>()
 		where TSpace : struct, IManagedBinaryBuffer<TSpace, T?> where T : struct
 		=> BufferManager.RegisterNullableSpace8<T, Composite<TSpace, TSpace, T?>>();
@@ -133,7 +148,8 @@ public static partial class BufferManager
 	/// Registers object space.
 	/// </summary>
 	/// <typeparam name="TSpace">Type of object space.</typeparam>
-	[ExcludeFromCodeCoverage, MethodImpl(MethodImplOptions.AggressiveInlining)]
+	[ExcludeFromCodeCoverage]
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static void RegisterSpace32<TSpace>() where TSpace : struct, IManagedBinaryBuffer<TSpace, Object>
 		=> BufferManager.RegisterSpace16<Composite<TSpace, TSpace, Object>>();
 	/// <summary>
@@ -141,7 +157,8 @@ public static partial class BufferManager
 	/// </summary>
 	/// <typeparam name="T">Type of items in the buffer.</typeparam>
 	/// <typeparam name="TSpace">Type of <typeparamref name="T"/> buffer.</typeparam>
-	[ExcludeFromCodeCoverage, MethodImpl(MethodImplOptions.AggressiveInlining)]
+	[ExcludeFromCodeCoverage]
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static void RegisterSpace32<T, TSpace>()
 		where TSpace : struct, IManagedBinaryBuffer<TSpace, T> where T : struct
 		=> BufferManager.RegisterSpace16<T, Composite<TSpace, TSpace, T>>();
@@ -150,7 +167,8 @@ public static partial class BufferManager
 	/// </summary>
 	/// <typeparam name="T">Type of nullable items in the space.</typeparam>
 	/// <typeparam name="TSpace">Type of <see name="Nullable{T}"/> space.</typeparam>
-	[ExcludeFromCodeCoverage, MethodImpl(MethodImplOptions.AggressiveInlining)]
+	[ExcludeFromCodeCoverage]
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static void RegisterNullableSpace32<T, TSpace>()
 		where TSpace : struct, IManagedBinaryBuffer<TSpace, T?> where T : struct
 		=> BufferManager.RegisterNullableSpace16<T, Composite<TSpace, TSpace, T?>>();
@@ -158,7 +176,8 @@ public static partial class BufferManager
 	/// Registers object space.
 	/// </summary>
 	/// <typeparam name="TSpace">Type of object space.</typeparam>
-	[ExcludeFromCodeCoverage, MethodImpl(MethodImplOptions.AggressiveInlining)]
+	[ExcludeFromCodeCoverage]
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static void RegisterSpace64<TSpace>() where TSpace : struct, IManagedBinaryBuffer<TSpace, Object>
 		=> BufferManager.RegisterSpace32<Composite<TSpace, TSpace, Object>>();
 	/// <summary>
@@ -166,7 +185,8 @@ public static partial class BufferManager
 	/// </summary>
 	/// <typeparam name="T">Type of items in the buffer.</typeparam>
 	/// <typeparam name="TSpace">Type of <typeparamref name="T"/> buffer.</typeparam>
-	[ExcludeFromCodeCoverage, MethodImpl(MethodImplOptions.AggressiveInlining)]
+	[ExcludeFromCodeCoverage]
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static void RegisterSpace64<T, TSpace>()
 		where TSpace : struct, IManagedBinaryBuffer<TSpace, T> where T : struct
 		=> BufferManager.RegisterSpace32<T, Composite<TSpace, TSpace, T>>();
@@ -175,7 +195,8 @@ public static partial class BufferManager
 	/// </summary>
 	/// <typeparam name="T">Type of nullable items in the space.</typeparam>
 	/// <typeparam name="TSpace">Type of <see name="Nullable{T}"/> space.</typeparam>
-	[ExcludeFromCodeCoverage, MethodImpl(MethodImplOptions.AggressiveInlining)]
+	[ExcludeFromCodeCoverage]
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static void RegisterNullableSpace64<T, TSpace>()
 		where TSpace : struct, IManagedBinaryBuffer<TSpace, T?> where T : struct
 		=> BufferManager.RegisterNullableSpace32<T, Composite<TSpace, TSpace, T?>>();
@@ -184,7 +205,8 @@ public static partial class BufferManager
 	/// Registers object space.
 	/// </summary>
 	/// <typeparam name="TSpace">Type of object space.</typeparam>
-	[ExcludeFromCodeCoverage, MethodImpl(MethodImplOptions.AggressiveInlining)]
+	[ExcludeFromCodeCoverage]
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static void RegisterSpace128<TSpace>() where TSpace : struct, IManagedBinaryBuffer<TSpace, Object>
 		=> BufferManager.RegisterSpace64<Composite<TSpace, TSpace, Object>>();
 	/// <summary>
@@ -192,7 +214,8 @@ public static partial class BufferManager
 	/// </summary>
 	/// <typeparam name="T">Type of items in the buffer.</typeparam>
 	/// <typeparam name="TSpace">Type of <typeparamref name="T"/> buffer.</typeparam>
-	[ExcludeFromCodeCoverage, MethodImpl(MethodImplOptions.AggressiveInlining)]
+	[ExcludeFromCodeCoverage]
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static void RegisterSpace128<T, TSpace>()
 		where TSpace : struct, IManagedBinaryBuffer<TSpace, T> where T : struct
 		=> BufferManager.RegisterSpace64<T, Composite<TSpace, TSpace, T>>();
@@ -201,7 +224,8 @@ public static partial class BufferManager
 	/// </summary>
 	/// <typeparam name="T">Type of nullable items in the space.</typeparam>
 	/// <typeparam name="TSpace">Type of <see name="Nullable{T}"/> space.</typeparam>
-	[ExcludeFromCodeCoverage, MethodImpl(MethodImplOptions.AggressiveInlining)]
+	[ExcludeFromCodeCoverage]
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static void RegisterNullableSpace128<T, TSpace>()
 		where TSpace : struct, IManagedBinaryBuffer<TSpace, T?> where T : struct
 		=> BufferManager.RegisterNullableSpace64<T, Composite<TSpace, TSpace, T?>>();
@@ -210,7 +234,8 @@ public static partial class BufferManager
 	/// Registers object space.
 	/// </summary>
 	/// <typeparam name="TSpace">Type of object space.</typeparam>
-	[ExcludeFromCodeCoverage, MethodImpl(MethodImplOptions.AggressiveInlining)]
+	[ExcludeFromCodeCoverage]
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static void RegisterSpace256<TSpace>() where TSpace : struct, IManagedBinaryBuffer<TSpace, Object>
 		=> BufferManager.RegisterSpace128<Composite<TSpace, TSpace, Object>>();
 	/// <summary>
@@ -218,7 +243,8 @@ public static partial class BufferManager
 	/// </summary>
 	/// <typeparam name="T">Type of items in the buffer.</typeparam>
 	/// <typeparam name="TSpace">Type of <typeparamref name="T"/> buffer.</typeparam>
-	[ExcludeFromCodeCoverage, MethodImpl(MethodImplOptions.AggressiveInlining)]
+	[ExcludeFromCodeCoverage]
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static void RegisterSpace256<T, TSpace>()
 		where TSpace : struct, IManagedBinaryBuffer<TSpace, T> where T : struct
 		=> BufferManager.RegisterSpace128<T, Composite<TSpace, TSpace, T>>();
@@ -227,7 +253,8 @@ public static partial class BufferManager
 	/// </summary>
 	/// <typeparam name="T">Type of nullable items in the space.</typeparam>
 	/// <typeparam name="TSpace">Type of <see name="Nullable{T}"/> space.</typeparam>
-	[ExcludeFromCodeCoverage, MethodImpl(MethodImplOptions.AggressiveInlining)]
+	[ExcludeFromCodeCoverage]
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static void RegisterNullableSpace256<T, TSpace>()
 		where TSpace : struct, IManagedBinaryBuffer<TSpace, T?> where T : struct
 		=> BufferManager.RegisterNullableSpace128<T, Composite<TSpace, TSpace, T?>>();
@@ -235,7 +262,8 @@ public static partial class BufferManager
 	/// Registers object space.
 	/// </summary>
 	/// <typeparam name="TSpace">Type of object space.</typeparam>
-	[ExcludeFromCodeCoverage, MethodImpl(MethodImplOptions.AggressiveInlining)]
+	[ExcludeFromCodeCoverage]
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static void RegisterSpace512<TSpace>() where TSpace : struct, IManagedBinaryBuffer<TSpace, Object>
 		=> BufferManager.RegisterSpace256<Composite<TSpace, TSpace, Object>>();
 	/// <summary>
@@ -243,7 +271,8 @@ public static partial class BufferManager
 	/// </summary>
 	/// <typeparam name="T">Type of items in the buffer.</typeparam>
 	/// <typeparam name="TSpace">Type of <typeparamref name="T"/> buffer.</typeparam>
-	[ExcludeFromCodeCoverage, MethodImpl(MethodImplOptions.AggressiveInlining)]
+	[ExcludeFromCodeCoverage]
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static void RegisterSpace512<T, TSpace>()
 		where TSpace : struct, IManagedBinaryBuffer<TSpace, T> where T : struct
 		=> BufferManager.RegisterSpace256<T, Composite<TSpace, TSpace, T>>();
@@ -252,7 +281,8 @@ public static partial class BufferManager
 	/// </summary>
 	/// <typeparam name="T">Type of nullable items in the space.</typeparam>
 	/// <typeparam name="TSpace">Type of <see name="Nullable{T}"/> space.</typeparam>
-	[ExcludeFromCodeCoverage, MethodImpl(MethodImplOptions.AggressiveInlining)]
+	[ExcludeFromCodeCoverage]
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static void RegisterNullableSpace512<T, TSpace>()
 		where TSpace : struct, IManagedBinaryBuffer<TSpace, T?> where T : struct
 		=> BufferManager.RegisterNullableSpace256<T, Composite<TSpace, TSpace, T?>>();
@@ -260,7 +290,8 @@ public static partial class BufferManager
 	/// Registers object space.
 	/// </summary>
 	/// <typeparam name="TSpace">Type of object space.</typeparam>
-	[ExcludeFromCodeCoverage, MethodImpl(MethodImplOptions.AggressiveInlining)]
+	[ExcludeFromCodeCoverage]
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static void RegisterSpace1024<TSpace>() where TSpace : struct, IManagedBinaryBuffer<TSpace, Object>
 		=> BufferManager.RegisterSpace512<Composite<TSpace, TSpace, Object>>();
 	/// <summary>
@@ -268,7 +299,8 @@ public static partial class BufferManager
 	/// </summary>
 	/// <typeparam name="T">Type of items in the buffer.</typeparam>
 	/// <typeparam name="TSpace">Type of <typeparamref name="T"/> buffer.</typeparam>
-	[ExcludeFromCodeCoverage, MethodImpl(MethodImplOptions.AggressiveInlining)]
+	[ExcludeFromCodeCoverage]
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static void RegisterSpace1024<T, TSpace>()
 		where TSpace : struct, IManagedBinaryBuffer<TSpace, T> where T : struct
 		=> BufferManager.RegisterSpace512<T, Composite<TSpace, TSpace, T>>();
@@ -277,7 +309,8 @@ public static partial class BufferManager
 	/// </summary>
 	/// <typeparam name="T">Type of nullable items in the space.</typeparam>
 	/// <typeparam name="TSpace">Type of <see name="Nullable{T}"/> space.</typeparam>
-	[ExcludeFromCodeCoverage, MethodImpl(MethodImplOptions.AggressiveInlining)]
+	[ExcludeFromCodeCoverage]
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static void RegisterNullableSpace1024<T, TSpace>()
 		where TSpace : struct, IManagedBinaryBuffer<TSpace, T?> where T : struct
 		=> BufferManager.RegisterNullableSpace512<T, Composite<TSpace, TSpace, T?>>();
