@@ -15,11 +15,8 @@ public partial class CString
 	{
 		ArgumentNullException.ThrowIfNull(value);
 		using BinaryConcatenator helper = new(separator);
-		for (Int32 index = 0; index < value.Length; index++)
-		{
-			CString? utf8Text = value[index];
+		foreach (CString? utf8Text in value.AsSpan())
 			helper.Write(utf8Text);
-		}
 		return helper.ToCString();
 	}
 	/// <summary>
@@ -139,11 +136,8 @@ public partial class CString
 	{
 		ArgumentNullException.ThrowIfNull(value);
 		using CStringConcatenator helper = new(separator);
-		for (Int32 index = 0; index < value.Length; index++)
-		{
-			CString? utf8Text = value[index];
+		foreach (CString? utf8Text in value.AsSpan())
 			helper.Write(utf8Text);
-		}
 		return helper.ToCString();
 	}
 	/// <summary>
