@@ -13,8 +13,6 @@ internal partial class FixedPointer
 		/// Internal <see cref="IDisposable"/> instance.
 		/// </summary>
 		private readonly IDisposable? _disposable;
-		/// <inheritdoc cref="Disposable{TFixed}.Value"/>
-		private readonly TFixed _fixed;
 
 		/// <summary>
 		/// Constructor.
@@ -23,7 +21,7 @@ internal partial class FixedPointer
 		/// <param name="disposable"><see cref="IDisposable"/> instance.</param>
 		protected Disposable(TFixed fixedPointer, IDisposable? disposable)
 		{
-			this._fixed = fixedPointer;
+			this.Value = fixedPointer;
 			this._disposable = disposable;
 		}
 
@@ -35,7 +33,7 @@ internal partial class FixedPointer
 		}
 
 		/// <inheritdoc/>
-		public TFixed Value => this._fixed;
+		public TFixed Value { get; }
 
 		~Disposable() { this.Dispose(false); }
 
@@ -50,8 +48,8 @@ internal partial class FixedPointer
 		/// <param name="disposing">Indicates whether current call is from <see cref="IDisposable.Dispose()"/>.</param>
 		private void Dispose(Boolean disposing)
 		{
-			if (!this._fixed.IsValid || (!disposing && this._disposable is IFixedPointer.IDisposable)) return;
-			this._fixed.Unload();
+			if (!this.Value.IsValid || (!disposing && this._disposable is IFixedPointer.IDisposable)) return;
+			this.Value.Unload();
 			this._disposable?.Dispose();
 		}
 	}
