@@ -6,7 +6,7 @@ internal unsafe partial class FixedPointer : IEquatable<FixedPointer>
 	/// <inheritdoc/>
 	public virtual Boolean Equals(FixedPointer? other)
 		=> other is not null && this.GetMemoryOffset() == other.GetMemoryOffset() &&
-			this.BinaryLength == other.BinaryLength && this._isReadOnly == other._isReadOnly;
+			this.BinaryLength == other.BinaryLength && this.IsReadOnly == other.IsReadOnly;
 	/// <inheritdoc/>
 	[ExcludeFromCodeCoverage]
 	public override Boolean Equals(Object? obj) => this.Equals(obj as FixedPointer);
@@ -19,7 +19,7 @@ internal unsafe partial class FixedPointer : IEquatable<FixedPointer>
 		result.Add(new IntPtr(this._ptr));
 		result.Add(this.BinaryOffset);
 		result.Add(this._binaryLength);
-		result.Add(this._isReadOnly);
+		result.Add(this.IsReadOnly);
 		if (this.Type is not null)
 			result.Add(this.Type);
 		return result.ToHashCode();
