@@ -10,7 +10,7 @@ public sealed class StringJoinEmptyTest
 	internal void LocalEmptyTest(Boolean emptyData)
 	{
 		String? separator = !emptyData ? String.Empty : default;
-		String?[] values = !emptyData ? Enumerable.Repeat(String.Empty, 3).ToArray() : Array.Empty<String>();
+		String?[] values = !emptyData ? Enumerable.Repeat(String.Empty, 3).ToArray() : [];
 		StringJoinEmptyTest.Test(separator, values);
 		StringJoinEmptyTest.Test(separator, values.ToList());
 	}
@@ -21,7 +21,7 @@ public sealed class StringJoinEmptyTest
 	internal void NullEmptyTest(Boolean emptyData)
 	{
 		String? separator = !emptyData ? String.Empty : default;
-		String?[] values = !emptyData ? Enumerable.Repeat<String?>(default, 3).ToArray() : Array.Empty<String>();
+		String?[] values = !emptyData ? Enumerable.Repeat<String?>(default, 3).ToArray() : [];
 		StringJoinEmptyTest.Test(separator, values);
 		StringJoinEmptyTest.Test(separator, values.ToList());
 	}
@@ -32,7 +32,7 @@ public sealed class StringJoinEmptyTest
 	internal async Task LocalEmptyTestAsync(Boolean emptyData)
 	{
 		String? separator = !emptyData ? String.Empty : default;
-		String?[] values = !emptyData ? Enumerable.Repeat(String.Empty, 3).ToArray() : Array.Empty<String>();
+		String?[] values = !emptyData ? Enumerable.Repeat(String.Empty, 3).ToArray() : [];
 		await StringJoinEmptyTest.TestAsync(separator, values);
 		await StringJoinEmptyTest.TestAsync(separator, values.ToList());
 	}
@@ -43,7 +43,7 @@ public sealed class StringJoinEmptyTest
 	internal async Task NullEmptyTestAsync(Boolean emptyData)
 	{
 		String? separator = !emptyData ? String.Empty : default;
-		String?[] values = !emptyData ? Enumerable.Repeat<String?>(default, 3).ToArray() : Array.Empty<String>();
+		String?[] values = !emptyData ? Enumerable.Repeat<String?>(default, 3).ToArray() : [];
 		await StringJoinEmptyTest.TestAsync(separator, values);
 		await StringJoinEmptyTest.TestAsync(separator, values.ToList());
 	}
