@@ -3,7 +3,7 @@ namespace Rxmxnx.PInvoke.Tests;
 [ExcludeFromCodeCoverage]
 internal sealed record TestMemoryHandle : IDisposable
 {
-	private readonly List<MemoryHandle> _handles = new();
+	private readonly List<MemoryHandle> _handles = [];
 	private Boolean _disposed;
 	public void Dispose()
 	{
